fixing default dictionary location; misc cleanup

+++ src/main/java/org/tinyradius/attribute/IntegerAttribute.java
@@ -75,7 +75,7 @@ public class IntegerAttribute extends RadiusAttribute {
         if (at != null) {
             Integer val = at.getEnumeration(value);
             if (val != null) {
-                setAttributeValue(val.intValue());
+                setAttributeValue(val);
                 return;
             }
         }
